--- a/src/uu/tr/Cargo.toml
+++ b/src/uu/tr/Cargo.toml
@@ -22,6 +22,7 @@ nom = { workspace = true }
 clap = { workspace = true }
 uucore = { workspace = true, features = ["fs"] }
 fluent = { workspace = true }
+bytecount = { workspace = true, features = ["runtime-dispatch-simd"] }
 
 [[bin]]
 name = "tr"