after #8439

With:
` dd if=/dev/zero bs=1M count=100 | tr '\0' 'x' > 100mb_input`

```
hyperfine --export-markdown tr.md --warmup 3 \
    '/usr/bin/tr a b < 100mb_input > /dev/null' \
    './target/release/tr a b < 100mb_input > /dev/null'     './target/release/tr.previous a b < 100mb_input > /dev/null'
Benchmark 1: /usr/bin/tr a b < 100mb_input > /dev/null
  Time (mean ± σ):      46.4 ms ±   5.4 ms    [User: 32.4 ms, System: 14.0 ms]
  Range (min … max):    40.2 ms …  63.9 ms    66 runs
 
Benchmark 2: ./target/release/tr a b < 100mb_input > /dev/null
  Time (mean ± σ):      29.3 ms ±   6.6 ms    [User: 14.0 ms, System: 15.2 ms]
  Range (min … max):    21.1 ms …  59.1 ms    91 runs
 
Benchmark 3: ./target/release/tr.previous a b < 100mb_input > /dev/null
  Time (mean ± σ):      85.0 ms ±   3.7 ms    [User: 71.9 ms, System: 13.0 ms]
  Range (min … max):    80.6 ms …  93.3 ms    34 runs
 
Summary
  ./target/release/tr a b < 100mb_input > /dev/null ran
    1.58 ± 0.40 times faster than /usr/bin/tr a b < 100mb_input > /dev/null
    2.90 ± 0.67 times faster than ./target/release/tr.previous a b < 100mb_input > /dev/null
```

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `/usr/bin/tr a b < 100mb_input > /dev/null` | 46.4 ± 5.4 | 40.2 | 63.9 | 1.58 ± 0.40 |
| `./target/release/tr a b < 100mb_input > /dev/null` | 29.3 ± 6.6 | 21.1 | 59.1 | 1.00 |
| `./target/release/tr.previous a b < 100mb_input > /dev/null` | 85.0 ± 3.7 | 80.6 | 93.3 | 2.90 ± 0.67 |